updated 'Known Bugs'
I will re-read how windows passes rotated frames and pass the rotation-mode to the functions. 
The rotation-mode should be obtainable from 'GetDesc()'. 
The picture gets passed in horizontal mode without a correct orientation. The obtained resolution for a, lets say vertical monitor (rotation mode 1 or so), is 1080x1920. We need to swap those values to not run out of bound while iterating over a virtual 1920x1080 matrix, which contains a side-ways lying frame.